# neploy/docs/swagger.yaml
basePath: /
definitions:
  model.Application:
    properties:
      appName:
        type: string
      createdAt:
        $ref: '#/definitions/model.Date'
      deletedAt:
        $ref: '#/definitions/model.Date'
      deployLocation:
        type: string
      description:
        type: string
      id:
        type: string
      storageLocation:
        type: string
      techStackId:
        type: string
      updatedAt:
        $ref: '#/definitions/model.Date'
    type: object
...
  model.CreateApplicationRequest:
    properties:
      appName:
        type: string
      description:
        type: string
    type: object
...
  model.Date:
    properties:
      time.Time:
        type: string
    type: object
...
  model.Gateway:
    properties:
      applicationId:
        type: string
      createdAt:
        $ref: '#/definitions/model.Date'
      deletedAt:
        $ref: '#/definitions/model.Date'
      domain:
        type: string
      endpointType:
        description: '"subdomain" or "path"'
        type: string
      endpointUrl:
        type: string
      httpMethod:
        type: string
      id:
        type: string
      integrationType:
        type: string
      loggingLevel:
        type: string
      name:
        type: string
      path:
        type: string
      port:
        type: string
      stage:
        type: string
      status:
        description: '"active", "inactive", "error"'
        type: string
      subdomain:
        type: string
      updatedAt:
        $ref: '#/definitions/model.Date'
    type: object
  model.GatewayConfig:
    properties:
      createdAt:
        $ref: '#/definitions/model.Date'
      defaultVersion:
        $ref: '#/definitions/model.VersionType'
      defaultVersioningType:
        $ref: '#/definitions/model.VersioningType'
      deletedAt:
        $ref: '#/definitions/model.Date'
      id:
        type: string
      loadBalancer:
        type: boolean
      updatedAt:
        $ref: '#/definitions/model.Date'
    type: object
...
  model.VersionType:
    enum:
    - latest
    - stable
    type: string
    x-enum-varnames:
    - VersionTypeLatest
    - VersionTypeStable
  model.VersioningType:
    enum:
    - header
    - uri
    type: string
    x-enum-varnames:
    - VersioningTypeHeader
    - VersioningTypeUri
host: localhost:8080
info:
  contact:
    email: 6oMwz@example.com
    name: API Support
    url: http://www.swagger.io/support
  description: Neploy API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Neploy API
  version: "1.0"
paths:
  /applications:
    get:
      consumes:
      - application/json
      description: List all applications
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Application'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all applications
      tags:
      - Application
    post:
      consumes:
      - application/json
      description: Create a new application
      parameters:
      - description: Application details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new application
      tags:
      - Application
...
  /gateways:
    post:
      consumes:
      - application/json
      description: Create a new gateway
      parameters:
      - description: Gateway details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Gateway'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new gateway
      tags:
      - Gateway
  /gateways/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a gateway
      parameters:
      - description: Gateway ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a gateway
      tags:
      - Gateway
    get:
      consumes:
      - application/json
      description: Get a gateway by ID
      parameters:
      - description: Gateway ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Gateway'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get a gateway by ID
      tags:
      - Gateway
    put:
      consumes:
      - application/json
      description: Update a gateway
      parameters:
      - description: Gateway ID
        in: path
        name: id
        required: true
        type: string
      - description: Gateway details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.Gateway'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a gateway
      tags:
      - Gateway
...
  /gateways/app/{appId}:
    get:
      consumes:
      - application/json
      description: List gateways by app ID
      parameters:
      - description: App ID
        in: path
        name: appId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Gateway'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List gateways by app ID
      tags:
      - Gateway
  /gateways/config:
    get:
      description: Gets the configuration of the API Gateway
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GatewayConfig'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get API Gateway Config
      tags:
      - Gateway
    post:
      consumes:
      - application/json
      description: Saves API Gateway Configurations like default versioning
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GatewayConfig'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Saves config
      tags:
      - Gateway
...
swagger: "2.0"